+++ sdk/attestation/azure-security-attestation/samples/policy-certificates/add_policy_certificate.cpp
@@ -50,9 +50,8 @@ int main()
         GetEnvHelper::GetEnv("AZURE_TENANT_ID"),
         GetEnvHelper::GetEnv("AZURE_CLIENT_ID"),
         GetEnvHelper::GetEnv("AZURE_CLIENT_SECRET"));
-    std::shared_ptr<AttestationAdministrationClient> adminClient(
-        AttestationAdministrationClient::CreatePointer(
-            GetEnvHelper::GetEnv("ATTESTATION_ISOLATED_URL"), credential));
+    AttestationAdministrationClient adminClient(AttestationAdministrationClient::Create(
+        GetEnvHelper::GetEnv("ATTESTATION_ISOLATED_URL"), credential));
 
     std::string const signingKey(GetEnvHelper::GetEnv("ISOLATED_SIGNING_KEY"));
     std::string const signingCert(GetEnvHelper::GetEnv("ISOLATED_SIGNING_CERTIFICATE"));
@@ -76,7 +75,7 @@ int main()
       // Add the new certificate to the set of policy management certificates for this attestation
       // service instance.
       Azure::Response<AttestationToken<IsolatedModeCertificateModificationResult>> const addResult
-          = adminClient->AddIsolatedModeCertificate(pemCertificateToAdd, requestSigner);
+          = adminClient.AddIsolatedModeCertificate(pemCertificateToAdd, requestSigner);
 
       std::cout << "The result of the certificate add operation is: "
                 << addResult.Value.Body.CertificateModification.ToString() << std::endl;
@@ -115,7 +114,7 @@ int main()
       // Add the new certificate to the set of policy management certificates for this attestation
       // service instance.
       Azure::Response<AttestationToken<IsolatedModeCertificateModificationResult>> const addResult
-          = adminClient->RemoveIsolatedModeCertificate(pemCertificateToRemove, requestSigner);
+          = adminClient.RemoveIsolatedModeCertificate(pemCertificateToRemove, requestSigner);
 
       std::cout << "The result of the certificate remove operation is: "
                 << addResult.Value.Body.CertificateModification.ToString() << std::endl;