+++ sdk/attestation/azure-security-attestation/samples/policy/set_policy.cpp
@@ -58,8 +58,8 @@ int main()
         GetEnvHelper::GetEnv("AZURE_TENANT_ID"),
         GetEnvHelper::GetEnv("AZURE_CLIENT_ID"),
         GetEnvHelper::GetEnv("AZURE_CLIENT_SECRET"));
-    std::unique_ptr<AttestationAdministrationClient const> adminClient(
-        AttestationAdministrationClient::CreatePointer(endpoint, credential, clientOptions));
+    AttestationAdministrationClient const adminClient(
+        AttestationAdministrationClient::Create(endpoint, credential, clientOptions));
 
     // Set the attestation policy on this attestation instance.
     // Note that because this is an AAD mode instance, the caller does not need to sign the policy
@@ -73,7 +73,7 @@ authorizationrules
 	[ type=="x-ms-sgx-mrsigner", value=="mrsigner2"] => permit(); 
 };)");
     Azure::Response<AttestationToken<PolicyResult>> const setResult
-        = adminClient->SetAttestationPolicy(AttestationType::SgxEnclave, policyToSet);
+        = adminClient.SetAttestationPolicy(AttestationType::SgxEnclave, policyToSet);
 
     if (setResult.Value.Body.PolicyResolution == PolicyModification::Updated)
     {
@@ -89,7 +89,7 @@ authorizationrules
     // by the attestation service, the customer can call CreateAttestationPolicyToken and then
     // generate the SHA256 of that token and compare it with the value returned by the service - the
     // two hash values should be identical.
-    auto const setPolicyToken = adminClient->CreateAttestationPolicyToken(policyToSet);
+    auto const setPolicyToken = adminClient.CreateAttestationPolicyToken(policyToSet);
     Sha256Hash shaHasher;
     std::vector<uint8_t> policyTokenHash = shaHasher.Final(
         reinterpret_cast<uint8_t const*>(setPolicyToken.RawToken.data()),