+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs.hpp
@@ -8,13 +8,18 @@
 
 #pragma once
 
+#include <azure/storage/common/storage_exception.hpp>
+
 #include "azure/storage/blobs/append_blob_client.hpp"
 #include "azure/storage/blobs/blob_client.hpp"
 #include "azure/storage/blobs/blob_container_client.hpp"
 #include "azure/storage/blobs/blob_lease_client.hpp"
+#include "azure/storage/blobs/blob_options.hpp"
+#include "azure/storage/blobs/blob_responses.hpp"
 #include "azure/storage/blobs/blob_sas_builder.hpp"
 #include "azure/storage/blobs/blob_service_client.hpp"
 #include "azure/storage/blobs/block_blob_client.hpp"
 #include "azure/storage/blobs/dll_import_export.hpp"
 #include "azure/storage/blobs/page_blob_client.hpp"
-#include "azure/storage/common/storage_exception.hpp"
+#include "azure/storage/blobs/rest_client.hpp"
+#include "azure/storage/blobs/rtti.hpp"