+++ sdk/storage/azure-storage-blobs/test/ut/block_blob_client_test.cpp
@@ -9,7 +9,6 @@
 
 #include <azure/core/cryptography/hash.hpp>
 #include <azure/storage/common/crypt.hpp>
-#include <azure/storage/common/internal/file_io.hpp>
 
 namespace Azure { namespace Storage { namespace Blobs { namespace Models {
 
@@ -191,10 +190,7 @@ namespace Azure { namespace Storage { namespace Test {
       }
       {
         const std::string tempFilename = "file" + testName;
-        {
-          Azure::Storage::_internal::FileWriter fileWriter(tempFilename);
-          fileWriter.Write(blobContent.data(), blobContent.size(), 0);
-        }
+        WriteFile(tempFilename, blobContent);
         client.UploadFrom(tempFilename, options);
         EXPECT_EQ(client.GetTags().Value, tags);
         client.Delete();
@@ -1122,9 +1118,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_NO_THROW(blockBlobClient.Delete());
 
     std::string emptyFilename(testName);
-    {
-      _internal::FileWriter writer(emptyFilename);
-    }
+    WriteFile(emptyFilename, std::vector<uint8_t>{});
     blockBlobClient.UploadFrom(emptyFilename);
     EXPECT_NO_THROW(blockBlobClient.Delete());
 
@@ -1281,7 +1275,7 @@ namespace Azure { namespace Storage { namespace Test {
     SetOptions();
     UploadBlockBlob::ParamType const& p(GetParam());
     auto const blobSize = p.Size;
-    std::vector<uint8_t> blobContent(static_cast<size_t>(8_MB), 'x');
+    std::vector<uint8_t> blobContent(static_cast<size_t>(p.Size), 'x');
 
     Azure::Storage::Blobs::UploadBlockBlobFromOptions options;
     options.TransferOptions.ChunkSize = 1_MB;
@@ -1292,10 +1286,7 @@ namespace Azure { namespace Storage { namespace Test {
     options.AccessTier = m_blobUploadOptions.AccessTier;
 
     std::string tempFilename(testName);
-    {
-      Azure::Storage::_internal::FileWriter fileWriter(tempFilename);
-      fileWriter.Write(blobContent.data(), static_cast<size_t>(blobSize), 0);
-    }
+    WriteFile(tempFilename, blobContent);
     auto res = blockBlobClient.UploadFrom(tempFilename, options);
     EXPECT_TRUE(res.Value.ETag.HasValue());
     EXPECT_TRUE(IsValidTime(res.Value.LastModified));